+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -14,7 +14,12 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.TopHeadlinesAppCompose"
-        tools:targetApi="31">
+        tools:targetApi="31" >
+        <activity
+            android:name=".presentation.topHeadlinesList.TopHeadlinesScreen"
+            android:exported="false"
+            android:label="@string/title_activity_top_headlines_screen"
+            android:theme="@style/Theme.TopHeadlinesAppCompose" />
         <activity
             android:name=".ui.DetailActivity"
             android:exported="false"
@@ -27,7 +32,7 @@
             android:name=".ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.TopHeadlinesAppCompose">
+            android:theme="@style/Theme.TopHeadlinesAppCompose" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
