+++ app/src/main/java/com/example/topheadlinesappcompose/ui/MainActivity.kt
@@ -43,15 +43,18 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.findNavController
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.example.topheadlinesappcompose.domain.model.TopHeadlinesItem
 import com.example.topheadlinesappcompose.presentation.navigation.TopHeadlinesScreens
+import com.example.topheadlinesappcompose.presentation.topHeadlinesDetails.TopHeadlinesDetailsScreen
 import com.example.topheadlinesappcompose.ui.theme.TopHeadlinesAppComposeTheme
 import com.example.topheadlinesappcompose.ui.theme.nunitoFontFamilyBold
 import com.example.topheadlinesappcompose.presentation.topHeadlinesList.TopHeadlineViewModel
+import com.example.topheadlinesappcompose.presentation.topHeadlinesList.TopHeadlinesListScreen
 import com.example.topheadlinesappcompose.utils.Resource
 import com.example.topheadlinesappcompose.utils.extensions.formatDate
 import dagger.hilt.android.AndroidEntryPoint
@@ -73,6 +76,12 @@ class MainActivity : ComponentActivity() {
                         NavHost(
                             navController = rememberNavController(),
                             startDestination = TopHeadlinesScreens.HeadLineList.name ){
+                            composable(TopHeadlinesScreens.HeadLineList.name){
+                                TopHeadlinesListScreen()
+                            }
+                            composable(TopHeadlinesScreens.HeadLineDetails.name){
+                                TopHeadlinesDetailsScreen()
+                            }
 
                         }
 
@@ -98,26 +107,26 @@ class MainActivity : ComponentActivity() {
         }
     }
 }
-@Composable
-fun setUpNavigation(){
-//    NavHost(navController = rememberNavController(), startDestination = "First"){
-//        composable("First"){
-//            FirstScreen()
-//        }
-//        composable("Second"){
-//            SecondScreen()
-//        }
-//    }
-}
+//@Composable
+//fun setUpNavigation(){
+////    NavHost(navController = rememberNavController(), startDestination = "First"){
+////        composable("First"){
+////            FirstScreen()
+////        }
+////        composable("Second"){
+////            SecondScreen()
+////        }
+////    }
+//}
 
 
-fun goToNextScreen(context: Context, topHeadlinesItem: TopHeadlinesItem?) {
-    val nextScreenIntent = Intent(context, DetailActivity::class.java)
-    nextScreenIntent.putExtra("data", topHeadlinesItem)
-    context.startActivity(nextScreenIntent)
-
-    //context.startActivity(Intent(context, DetailActivity::class.java))
-}
+//fun goToNextScreen(context: Context, topHeadlinesItem: TopHeadlinesItem?) {
+//    val nextScreenIntent = Intent(context, DetailActivity::class.java)
+//    nextScreenIntent.putExtra("data", topHeadlinesItem)
+//    context.startActivity(nextScreenIntent)
+//
+//    //context.startActivity(Intent(context, DetailActivity::class.java))
+//}
 //@Composable
 //fun buttonClick(){
 //    val context = LocalContext.current
@@ -129,124 +138,125 @@ fun goToNextScreen(context: Context, topHeadlinesItem: TopHeadlinesItem?) {
 //    }
 //}
 
-@Composable
-fun getTopHeadlines() {
- val topViewModel : TopHeadlineViewModel = hiltViewModel()
-//    val lifecycle = LocalLifecycleOwner.current
-//    LaunchedEffect(Unit){
-//        lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED){
-//            topViewModel.getTopHeadlines("bbc-news").collect{
-//                when(it){
-//                    is Resource.Error -> println("ERROR:: ${it.error}")
-//                    is Resource.Success -> println("SUCCESS::: ${it.data}")
-//                }
+//@Composable
+//fun getTopHeadlines() {
+// val topViewModel : TopHeadlineViewModel = hiltViewModel()
+////    val lifecycle = LocalLifecycleOwner.current
+////    LaunchedEffect(Unit){
+////        lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED){
+////            topViewModel.getTopHeadlines("bbc-news").collect{
+////                when(it){
+////                    is Resource.Error -> println("ERROR:: ${it.error}")
+////                    is Resource.Success -> println("SUCCESS::: ${it.data}")
+////                }
+////            }
+////        }
+////    }
+////    val mutableState = remember {
+////        TopHeadlinesResponse()
+////    }
+//
+//
+//    val response by topViewModel.getTopHeadlines.collectAsStateWithLifecycle(Resource.Loading)
+//    when(response){
+//        is Resource.Error -> print("ERROR:::${response.error}")
+//        Resource.Loading -> println("Loading")
+//        is Resource.Success -> {
+//            println("SUCCESS::: ${response.data}")
+//            val data = response.data
+//                TopHeadLineList(topHeadlinesItem = data)
+//        }
+//    }
+//
+//}
+
+//@Composable
+//fun TopHeadLineList(topHeadlinesItem: List<TopHeadlinesItem>?){
+//    LazyColumn(modifier = Modifier.fillMaxSize())
+//    {
+//        topHeadlinesItem?.size?.let {
+//            items(it) { i ->
+//                TopHeadLinesItemComposable(topHeadlinesItem = topHeadlinesItem[i])
+//
 //            }
 //        }
 //    }
-//    val mutableState = remember {
-//        TopHeadlinesResponse()
-//    }
+//
+//}
 
-
-    val response by topViewModel.getTopHeadlines.collectAsStateWithLifecycle(Resource.Loading)
-    when(response){
-        is Resource.Error -> print("ERROR:::${response.error}")
-        Resource.Loading -> println("Loading")
-        is Resource.Success -> {
-            println("SUCCESS::: ${response.data}")
-            val data = response.data
-                TopHeadLineList(topHeadlinesItem = data)
-        }
-    }
-
-}
-@Composable
-fun TopHeadLineList(topHeadlinesItem: List<TopHeadlinesItem>?){
-    LazyColumn(modifier = Modifier.fillMaxSize())
-    {
-         items(10) {i ->
-             TopHeadLinesItemComposable(topHeadlinesItem = topHeadlinesItem?.get(i))
-
-         }
-    }
-
-}
-
-@OptIn(ExperimentalGlideComposeApi::class)
-@Composable
-fun TopHeadLinesItemComposable(topHeadlinesItem: TopHeadlinesItem?){
-    val context = LocalContext.current
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 7.dp, vertical = 10.dp)
-            .clickable {
-                goToNextScreen(context, topHeadlinesItem)
-            },
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-        colors = CardDefaults.cardColors(
-            containerColor = Color.White
-        ),
-        shape = RoundedCornerShape(corner = CornerSize(5.dp))
-
-
-    ) {
-   Row {
-       Box(modifier = Modifier.fillMaxWidth(),
-
-
-       ){
-           GlideImage(
-               model = topHeadlinesItem?.imageUrl,
-               contentDescription = "background Image",
-               contentScale = ContentScale.FillBounds,
-               modifier = Modifier
-                   .height(200.dp)
-                   .fillMaxWidth()
-
-               )
-//           Image(
-//               painter = painterResource(id = R.drawable.wall),
-//               contentDescription = "background",
+//@OptIn(ExperimentalGlideComposeApi::class)
+//@Composable
+//fun TopHeadLinesItemComposable(topHeadlinesItem: TopHeadlinesItem?){
+//    val context = LocalContext.current
+//    Card(
+//        modifier = Modifier
+//            .fillMaxWidth()
+//            .padding(horizontal = 7.dp, vertical = 10.dp)
+//            .clickable {
+//                goToNextScreen(context, topHeadlinesItem)
+//            },
+//        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
+//        colors = CardDefaults.cardColors(
+//            containerColor = Color.White
+//        ),
+//        shape = RoundedCornerShape(corner = CornerSize(5.dp))
+//
+//
+//    ) {
+//   Row {
+//       Box(modifier = Modifier.fillMaxWidth(),
+//       ){
+//           GlideImage(
+//               model = topHeadlinesItem?.imageUrl,
+//               contentDescription = "background Image",
 //               contentScale = ContentScale.FillBounds,
 //               modifier = Modifier
 //                   .height(200.dp)
 //                   .fillMaxWidth()
-//                   //.matchParentSize()
-//           )
-           Column(
-               modifier = Modifier
-                   .padding(16.dp)
-                   .fillMaxWidth()
-                   .align(Alignment.BottomEnd),
-               verticalArrangement = Arrangement.Bottom,
-               ) {
-               Text(text = topHeadlinesItem?.title.toString(),
-                   color = Color.White,
-                   fontSize = 18.sp,
-                   fontWeight = FontWeight.Bold,
-                   fontFamily = nunitoFontFamilyBold
-               )
-               Spacer(modifier = Modifier.size(5.dp))
-               Text(
-                   modifier = Modifier.fillMaxWidth(),
-                   text = topHeadlinesItem?.publishedAt?.formatDate().toString(),
-                   color = Color.White,
-                   fontWeight = FontWeight.Bold,
-                   fontFamily = nunitoFontFamilyBold,
-                   fontSize = 12.sp,
-                   textAlign = TextAlign.End
-               )
-           }
-       }
-
-   }
-
-
-    }
-
-
-}
+//
+//               )
+////           Image(
+////               painter = painterResource(id = R.drawable.wall),
+////               contentDescription = "background",
+////               contentScale = ContentScale.FillBounds,
+////               modifier = Modifier
+////                   .height(200.dp)
+////                   .fillMaxWidth()
+////                   //.matchParentSize()
+////           )
+//           Column(
+//               modifier = Modifier
+//                   .padding(16.dp)
+//                   .fillMaxWidth()
+//                   .align(Alignment.BottomEnd),
+//               verticalArrangement = Arrangement.Bottom,
+//               ) {
+//               Text(text = topHeadlinesItem?.title.toString(),
+//                   color = Color.White,
+//                   fontSize = 18.sp,
+//                   fontWeight = FontWeight.Bold,
+//                   fontFamily = nunitoFontFamilyBold
+//               )
+//               Spacer(modifier = Modifier.size(5.dp))
+//               Text(
+//                   modifier = Modifier.fillMaxWidth(),
+//                   text = topHeadlinesItem?.publishedAt?.formatDate().toString(),
+//                   color = Color.White,
+//                   fontWeight = FontWeight.Bold,
+//                   fontFamily = nunitoFontFamilyBold,
+//                   fontSize = 12.sp,
+//                   textAlign = TextAlign.End
+//               )
+//           }
+//       }
+//
+//   }
+//
+//
+//    }
+//
+//
+//}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -273,13 +283,13 @@ fun TopAppBar(){
 @Composable
 fun GreetingPreview() {
     TopHeadlinesAppComposeTheme {
-       TopHeadLinesItemComposable(topHeadlinesItem = TopHeadlinesItem(
-           "Birmingham sack Rooney after 15 games in charge",
-           "Birmingham sack Rooney after 15 games in charge",
-           "",
-           "",
-           "2024-01-02T12:22:23.4088451Z",
-           "") )
+//       TopHeadLinesItemComposable(topHeadlinesItem = TopHeadlinesItem(
+//           "Birmingham sack Rooney after 15 games in charge",
+//           "Birmingham sack Rooney after 15 games in charge",
+//           "",
+//           "",
+//           "2024-01-02T12:22:23.4088451Z",
+//           "") )
     }
