+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">TopHeadlinesAppCompose</string>
     <string name="title_activity_detail">DetailActivity</string>
+    <string name="title_activity_top_headlines_screen">TopHeadlinesScreen</string>
 </resources>
